requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.2
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
click==8.1.7
rich==13.7.0 